FIT: add seeds for fuzzing fitimage signing
To test the fitimage signing codepath a fitimage with a signing key is
needed as input. The file sandbox.its describes this fitimage, which
than can be generated with mkimage.
Three different iterations from this input are generated:
    fit/fit-image-null.bin
    fit/fit-image-only-dtb.bin
    fit/fit-image-sign.bin
These represent a fitimage with no data for kernel and dtb, only a dtb
and with a dtb and a signing key. These inputs result in different code
paths in the fitimage parser.
The binary seeds that are generated from these inputs by the fuzzer are
also added.

Signed-off-by: Steffen Trumtrar <s.trumtrar@pengutronix.de>